+++ cmd/cronosd/cmd/fix-unlucky-tx.go
@@ -111,14 +111,17 @@ func FixUnluckyTxCmd() *cobra.Command {
 							break
 						}
 
+						var hashes []string
 						txIndex++
 						for msgIndex, msg := range tx.GetMsgs() {
 							ethTxIndex := txIndex + int64(msgIndex)
 							ethTx := msg.(*evmtypes.MsgEthereumTx)
+							txHash := ethTx.AsTransaction().Hash().Hex()
+							hashes = append(hashes, txHash)
 							evt := abci.Event{
 								Type: evmtypes.TypeMsgEthereumTx,
 								Attributes: []abci.EventAttribute{
-									{Key: []byte(evmtypes.AttributeKeyEthereumTxHash), Value: []byte(ethTx.Hash), Index: true},
+									{Key: []byte(evmtypes.AttributeKeyEthereumTxHash), Value: []byte(txHash), Index: true},
 									{Key: []byte(evmtypes.AttributeKeyTxIndex), Value: []byte(strconv.FormatInt(ethTxIndex, 10)), Index: true},
 								},
 							}
@@ -140,8 +143,8 @@ func FixUnluckyTxCmd() *cobra.Command {
 						}); err != nil {
 							return err
 						}
-						for _, msg := range tx.GetMsgs() {
-							fmt.Println("patched", height, msg.(*evmtypes.MsgEthereumTx).Hash)
+						for _, txHash := range hashes {
+							fmt.Println("patched", height, txHash)
 						}
 						return nil
 					} else if txResult.Code == 0 {